Re-structured code implementation, need to fix mem allocation
