+++ main_alpakaTBB.cc
@@ -6,14 +6,7 @@
 #include "input.h"
 #include "output.h"
 
-
-/* Do NOT include other headers that use CUDA runtime functions or variables
- * before this include, because cupla renames CUDA host functions and device
- * built-in variables using macros and macro functions.
- * Do NOT include other specific includes such as `<cuda.h>`, etc.
- */
-
-#include "rawtodigi_alpakaSerial.h"
+#include "rawtodigi_alpakaTBB.h"
 
 namespace {
   constexpr int NLOOPS = 100;
@@ -48,14 +41,17 @@ int main(){
     std::cout << "Got " << input.cablingMap.size << " for cabling, wordCounter " << input.wordCounter << std::endl;
     
     //Selecting the device to execute on
-    using CpuParallelTBB = CpuParallelTBB<1u>;
-    using Idx = CpuParallelTBB::Idx;
+    // using CpuSerial = CpuSerial<1u>;
+    // using Idx = CpuSerial::Idx;
+    using namespace CPU_TBB;
 
-    CpuParallelTBB::DevHost const devHost(alpaka::pltf::getDevByIdx<CpuParallelTBB::PltfHost>(0u));
-    CpuParallelTBB::DevAcc const devAcc(alpaka::pltf::getDevByIdx<CpuParallelTBB::PltfAcc>(0u));
+    DevHost const devHost(alpaka::pltf::getDevByIdx<PltfHost>(0u));
+    DevAcc const devAcc(alpaka::pltf::getDevByIdx<PltfAcc>(0u));
+    Idx const elements(1);
+    Vec const extent(elements);
 
-    //Creating a queue on the device
-    CpuParallelTBB::QueueSync queue(devAcc); //Synchronus
+    // //Creating a queue on the device
+    QueueSync queue(devAcc); //Synchronus
      
     int totaltime = 0;
 
@@ -65,14 +61,14 @@ int main(){
     
     output = std::make_unique<Output>();
 
-    auto const input_dBuf(alpaka::mem::buf::alloc<Input, Idx>(devAcc, sizeof(Input)));
-    auto const input_hBuf(alpaka::mem::buf::alloc<Input, Idx>(devHost, sizeof(Input)));
+    auto const input_dBuf(alpaka::mem::buf::alloc<Input, Idx>(devAcc, extent ));
+    auto const input_hBuf(alpaka::mem::buf::alloc<Input, Idx>(devHost, extent));
 
     auto* pinput_dBuf(alpaka::mem::view::getPtrNative(input_dBuf));
     pinput_dBuf = &input;
 
-    auto output_dBuf(alpaka::mem::buf::alloc<Output, Idx>(devAcc, sizeof(Output)));
-    auto const output_hBuf(alpaka::mem::buf::alloc<Output, Idx>(devHost, sizeof(Output)));
+    auto output_dBuf(alpaka::mem::buf::alloc<Output, Idx>(devAcc, extent));
+    auto const output_hBuf(alpaka::mem::buf::alloc<Output, Idx>(devHost, extent));
 
     auto* poutput_dBuf(alpaka::mem::view::getPtrNative(output_dBuf));
 
@@ -81,9 +77,9 @@ int main(){
     // std::cout << pinput_dBuf << std::endl;
     auto start = std::chrono::high_resolution_clock::now();
 
-    alpaka::rawtodigi(pinput_dBuf, poutput_dBuf, input.wordCounter, false, false, false, queue);
+    Alpaka::rawtodigi(pinput_dBuf, poutput_dBuf, input.wordCounter, true, true, true, queue);
 
-    //Sync queue?
+    // //Sync queue?
 
     auto stop = std::chrono::high_resolution_clock::now();
