+++ output_alpaka.h
@@ -1,16 +1,12 @@
-#ifndef OUTPUT_H
-#define OUTPUT_H
+#ifndef OUTPUT_ALPAKA_H
+#define OUTPUT_ALPAKA_H
 
 #include <vector>
 
 #include "pixelgpudetails.h"
 
-#if defined DIGI_CUDA || defined DIGI_CUPLA || defined DIGI_KOKKOS 
-#include "GPUSimpleVector.h"
-
-#elif defined DIGI_ALPAKA
 #include "GPUSimpleVector_alpaka.h"
-#endif
+
 
 struct alignas(128) Output {
   uint16_t xx[pixelgpudetails::MAX_FED_WORDS];
@@ -21,12 +17,8 @@ struct alignas(128) Output {
   uint32_t digi[pixelgpudetails::MAX_FED_WORDS];
   uint32_t rawIdArr[pixelgpudetails::MAX_FED_WORDS];
 
-#ifdef DIGI_NAIVE
-  std::vector<PixelErrorCompact> err;
-#elif defined DIGI_CUDA || defined DIGI_CUPLA || defined DIGI_KOKKOS || defined DIGI_ALPAKA
   PixelErrorCompact err_d[pixelgpudetails::MAX_FED_WORDS];
-  GPU::SimpleVector<PixelErrorCompact> err;
-#endif
+
 };
 
 #endif
