+++ rawtodigi_alpakaGpu.h
@@ -1,24 +1,27 @@
 #ifndef RAWTODIGI_ALPAKAGPU_H
 #define RAWTODIGI_ALPAKAGPU_H
 
-/* Do NOT include other headers that use CUDA runtime functions or variables
- * before this include, because cupla renames CUDA host functions and device
- * built-in variables using macros and macro functions.
- * Do NOT include other specific includes such as `<cuda.h>`, etc.
- */
+#define ALPAKA_ARCHITECTURE GPU_CUDA
 
 #include "alpakaConfigGpu.h"
-#include "pixelgpudetails.h"
-#include "GPUSimpleVector.h"
-#include "input.h"
-#include "output.h"
 
+namespace GPU_CUDA{
 
-namespace alpaka {
-  
-  void rawtodigi(const Input *input_d, Output *output_d,
-                const uint32_t wordCounter,
-                bool useQualityInfo, bool includeErrors, bool debug, GpuCuda<1u>::QueueAsync queue);
-}
+    using Accelerator = GpuCuda<1u>;
+    using Acc = Accelerator::Acc;
+    using Dim = Accelerator::Dim;
+    using Idx = Accelerator::Idx;
+    using DevHost = Accelerator::DevHost;
+    using DevAcc = Accelerator::DevAcc;
+    using PltfHost = Accelerator::PltfHost;
+    using PltfAcc = Accelerator::PltfAcc;
+    using Extent = Accelerator::Extent;
+    using QueueAsync = Accelerator::QueueAsync;
+    using QueueSync = Accelerator::QueueSync;
+    using Vec = Accelerator::Vec;
+    using WorkDiv = Accelerator::WorkDiv;
+};
 
+#include "rawtodigi_alpaka.h"
+// #undef ALPAKA_ARCHITECTURE
 #endif
