+++ rawtodigi_alpakaSer.cc
@@ -1,0 +1,5 @@
+#include "rawtodigi_alpakaSer.h"
+
+#define ALPAKA_ARCHITECTURE CPU_SERIAL
+#include "rawtodigi_alpaka.cc"
+// #undef ALPAKA_ARCHITECTURE
